requirements.txt
python-magic
numpy
scipy
Pillow
